+++ packages/codegen/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Breaking
+
+- Decouple descriptor types from descriptor values
+
 ## 0.4.0 - 2024-05-03
 
 ### Breaking