+++ packages/codegen/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.5.0 - 2024-05-10
+
 ### Breaking
 
 - Decouple descriptor types from descriptor values